tidies routes a bit to make less daunting for Darren
