+++ app/config.js
@@ -14,6 +14,21 @@ module.exports = {
   useAuth: 'true',
 
   // Cookie warning - update link to service's cookie page.
-  cookieText: 'GOV.UK uses cookies to make the site simpler. <a href="#" title="Find out more about cookies">Find out more about cookies</a>'
+  cookieText: 'GOV.UK uses cookies to make the site simpler. <a href="#" title="Find out more about cookies">Find out more about cookies</a>',
+
+  // this could be accessed from a variety of required js modules so placed here in config
+  protoPaths: {
+    version: '/:phase/:version*',                     // e.g '/alpha/alpha-01/'
+    step: '/:phase/:version*/app/:step',              // e.g '/alpha/alpha-01/app/address'
+    appsGlob: [
+      __dirname + '/views/**/index.html',
+      '!' + __dirname + '/views/index.html',
+      '!' + __dirname + '/views/**/app/index.html',
+      '!' + __dirname + 'views/includes/**/.*'
+    ],
+    routesGlob: [
+      __dirname + '/views/**/version_routes.js'
+    ]
+  }
 
 };
