+++ app/views/alpha/alpha-07/version_routes.js
@@ -1,5 +1,5 @@
 module.exports = function(router, config) {
-  router.all(config.protoPaths.step, function(req,res,next){
+  router.all(config.route, function(req,res,next){
 
     var requestedPage = req.params.step,
         postData = req.body || {};
